Get and expose backend config for preferred-server
Gets the CE_IPSEC_PREFERREDSERVER environment variable
and exposes this bool as part of the backend config.
