+++ pkg/util/util.go
@@ -19,16 +19,11 @@ import (
 	"fmt"
 	"log"
 	"net"
-	"os"
 	"strings"
 	"syscall"
 
-	"github.com/pkg/errors"
-	"github.com/rdegges/go-ipify"
 	level "github.com/submariner-io/admiral/pkg/log"
-	"github.com/submariner-io/admiral/pkg/stringset"
 	"github.com/vishvananda/netlink"
-	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/equality"
 	"k8s.io/klog"
 
@@ -106,122 +101,6 @@ func FlattenColors(colorCodes []string) string {
 	}
 
 	return flattenedColors
-}
-
-func GetLocalEndpoint(submSpec types.SubmarinerSpecification, privateIP string, gwNode *v1.Node) (types.SubmarinerEndpoint, error) {
-	hostname, err := os.Hostname()
-	if err != nil {
-		return types.SubmarinerEndpoint{}, fmt.Errorf("error getting hostname: %v", err)
-	}
-
-	var localSubnets []string
-
-	globalnetEnabled := false
-
-	if len(submSpec.GlobalCidr) > 0 {
-		localSubnets = submSpec.GlobalCidr
-		globalnetEnabled = true
-	} else {
-		localSubnets = append(localSubnets, submSpec.ServiceCidr...)
-		localSubnets = append(localSubnets, submSpec.ClusterCidr...)
-	}
-
-	backendConfig, err := getBackendConfig(gwNode)
-	if err != nil {
-		return types.SubmarinerEndpoint{}, err
-	}
-
-	endpoint := types.SubmarinerEndpoint{
-		Spec: subv1.EndpointSpec{
-			CableName:     fmt.Sprintf("submariner-cable-%s-%s", submSpec.ClusterID, strings.ReplaceAll(privateIP, ".", "-")),
-			ClusterID:     submSpec.ClusterID,
-			Hostname:      hostname,
-			PrivateIP:     privateIP,
-			NATEnabled:    submSpec.NatEnabled,
-			Subnets:       localSubnets,
-			Backend:       submSpec.CableDriver,
-			BackendConfig: backendConfig,
-		},
-	}
-
-	publicIP, err := ipify.GetIp()
-	if err != nil {
-		return types.SubmarinerEndpoint{}, fmt.Errorf("could not determine public IP: %v", err)
-	}
-
-	endpoint.Spec.PublicIP = publicIP
-
-	if !globalnetEnabled {
-		// When globalnet is enabled, HealthCheckIP will be the globalIP assigned to the Active GatewayNode.
-		// In a fresh deployment, globalIP annotation for the node might take few seconds. So we listen on NodeEvents
-		// and update the endpoint HealthCheckIP (to globalIP) in datastoreSyncer at a later stage. This will trigger
-		// the HealthCheck between the clusters.
-		endpoint.Spec.HealthCheckIP, err = getCNIInterfaceIPAddress(submSpec.ClusterCidr)
-		if err != nil {
-			return types.SubmarinerEndpoint{}, fmt.Errorf("error getting CNI Interface IP address: %v", err)
-		}
-	}
-
-	return endpoint, nil
-}
-
-func getBackendConfig(nodeObj *v1.Node) (map[string]string, error) {
-	validConfigs := stringset.New(subv1.ValidGatewayNodeConfig...)
-	backendConfig := map[string]string{}
-
-	for label, value := range nodeObj.Labels {
-		if strings.HasPrefix(label, subv1.GatewayConfigLabelPrefix) {
-			config := label[len(subv1.GatewayConfigLabelPrefix):]
-			if !validConfigs.Contains(config) {
-				return backendConfig, errors.Errorf("unknown config label %q on node %q", nodeObj.Name, label)
-			}
-
-			backendConfig[config] = value
-		}
-	}
-
-	return backendConfig, nil
-}
-
-//TODO: to handle de-duplication of code/finding common parts with the route agent
-func getCNIInterfaceIPAddress(clusterCIDRs []string) (string, error) {
-	for _, clusterCIDR := range clusterCIDRs {
-		_, clusterNetwork, err := net.ParseCIDR(clusterCIDR)
-		if err != nil {
-			return "", fmt.Errorf("unable to ParseCIDR %q : %v", clusterCIDR, err)
-		}
-
-		hostInterfaces, err := net.Interfaces()
-		if err != nil {
-			return "", fmt.Errorf("net.Interfaces() returned error : %v", err)
-		}
-
-		for _, iface := range hostInterfaces {
-			addrs, err := iface.Addrs()
-			if err != nil {
-				return "", fmt.Errorf("for interface %q, iface.Addrs returned error: %v", iface.Name, err)
-			}
-
-			for i := range addrs {
-				ipAddr, _, err := net.ParseCIDR(addrs[i].String())
-				if err != nil {
-					klog.Errorf("Unable to ParseCIDR : %q", addrs[i].String())
-				} else if ipAddr.To4() != nil {
-					klog.V(level.DEBUG).Infof("Interface %q has %q address", iface.Name, ipAddr)
-					address := net.ParseIP(ipAddr.String())
-
-					// Verify that interface has an address from cluster CIDR
-					if clusterNetwork.Contains(address) {
-						klog.V(level.DEBUG).Infof("Found CNI Interface %q that has IP %q from ClusterCIDR %q",
-							iface.Name, ipAddr.String(), clusterCIDR)
-						return ipAddr.String(), nil
-					}
-				}
-			}
-		}
-	}
-
-	return "", fmt.Errorf("unable to find CNI Interface on the host which has IP from %q", clusterCIDRs)
 }
 
 func GetClusterIDFromCableName(cableName string) string {
