+++ pkg/util/util_test.go
@@ -18,8 +18,6 @@ package util_test
 import (
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-	v1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	subv1 "github.com/submariner-io/submariner/pkg/apis/submariner.io/v1"
 	"github.com/submariner-io/submariner/pkg/types"
@@ -30,8 +28,6 @@ var _ = Describe("Util", func() {
 	Describe("Function ParseSecure", testParseSecure)
 
 	Describe("Function FlattenColors", testFlattenColors)
-
-	Describe("Function GetLocalEndpoint", testGetLocalEndpoint)
 
 	Describe("Function GetClusterIDFromCableName", testGetClusterIDFromCableName)
 
@@ -79,57 +75,6 @@ func testFlattenColors() {
 	Context("with no elements", func() {
 		It("should return an empty string", func() {
 			Expect(util.FlattenColors([]string{})).To(Equal(""))
-		})
-	})
-}
-
-func testGetLocalEndpoint() {
-	var submSpec types.SubmarinerSpecification
-	var node *v1.Node
-	const (
-		testPrivateIP       = "127.0.0.1"
-		testUDPPort         = "1111"
-		testUDPPortLabel    = "udp-port"
-		testNATTPortLabel   = "natt-discovery-port"
-		backendConfigPrefix = "gateway.submariner.io/"
-	)
-
-	subnets := []string{"127.0.0.1/16"}
-
-	BeforeEach(func() {
-		submSpec = types.SubmarinerSpecification{
-			ClusterID:   "east",
-			ClusterCidr: subnets,
-			CableDriver: "backend",
-		}
-
-		node = &v1.Node{
-			ObjectMeta: metav1.ObjectMeta{
-				Labels: map[string]string{
-					backendConfigPrefix + testNATTPortLabel: "1234",
-					backendConfigPrefix + testUDPPortLabel:  testUDPPort,
-				}}}
-	})
-
-	It("should return a valid SubmarinerEndpoint object", func() {
-		endpoint, err := util.GetLocalEndpoint(submSpec, testPrivateIP, node)
-
-		Expect(err).ToNot(HaveOccurred())
-		Expect(endpoint.Spec.ClusterID).To(Equal("east"))
-		Expect(endpoint.Spec.CableName).To(HavePrefix("submariner-cable-east-"))
-		Expect(endpoint.Spec.Hostname).NotTo(Equal(""))
-		Expect(endpoint.Spec.PrivateIP).To(Equal(testPrivateIP))
-		Expect(endpoint.Spec.Backend).To(Equal("backend"))
-		Expect(endpoint.Spec.Subnets).To(Equal(subnets))
-		Expect(endpoint.Spec.NATEnabled).To(Equal(false))
-		Expect(endpoint.Spec.BackendConfig[testUDPPortLabel]).To(Equal(testUDPPort))
-	})
-
-	When("no NAT discovery port label is set on the node", func() {
-		It("should return a valid SubmarinerEndpoint object", func() {
-			delete(node.Labels, testNATTPortLabel)
-			_, err := util.GetLocalEndpoint(submSpec, testPrivateIP, node)
-			Expect(err).ToNot(HaveOccurred())
 		})
 	})
 }
